added two attributes to issues and updated issues.dtd Namely
ballot
This is the ballot in which the issues was raised (if it was) e.g. "PLCS DIS ballot"

ballot_comment_id
This is the ID for the issue in the ballot. If the ballot issue results in a SEDS, the issue ID will become the SEDS id, therefore, may get renumbered to. The
ballot_comment_id is used to retain the original ballot comment id for auditing purposes.