changed mapping_view_with_test.xml with mapping_view.xml. email from Mike Ward today. Hi again
Just had a word with Nigel who reminded me what the score was with “mapping view” and “mapping view with test”…..

These were independent initiatives.

“mapping view” has more tests but works from the tree of short forms
“mapping view with test” constructs a pseudo (MIM) long-form and works from that

It should be noted that “mapping view with test” was not completed and seems to be a bit broken. It is constructing some faulty paths and generating lots of spurious errors. These seem to arise from the fact it has not been designed to cope with the newer ( /MAPPING_OF ( <ARM_element> ) / ) syntax. The correct errors it is finding – basically missing select type extensions in the MIM – are also found by “mapping view”.

Mike